Move VxScan backend CVR export related tests into their own file
