+++ apps/scan/backend/src/app_config.test.ts
@@ -7,32 +7,18 @@ import waitForExpect from 'wait-for-expect';
 import { LogEventId } from '@votingworks/logging';
 import {
   BooleanEnvironmentVariableName,
-  convertCastVoteRecordVotesToTabulationVotes,
-  getCurrentSnapshot,
   getFeatureFlagMock,
   singlePrecinctSelectionFor,
 } from '@votingworks/utils';
-import {
-  assertDefined,
-  err,
-  find,
-  iter,
-  ok,
-  typedAs,
-  unique,
-} from '@votingworks/basics';
+import { err, find, iter, ok, typedAs, unique } from '@votingworks/basics';
 import {
   fakeElectionManagerUser,
   fakeSessionExpiresAt,
   mockOf,
 } from '@votingworks/test-utils';
-import {
-  getCastVoteRecordExportDirectoryPaths,
-  mockElectionPackageFileTree,
-  readCastVoteRecordExport,
-} from '@votingworks/backend';
+import { mockElectionPackageFileTree } from '@votingworks/backend';
 import { InsertedSmartCardAuthApi } from '@votingworks/auth';
-import { CVR, ElectionDefinition } from '@votingworks/types';
+import { ElectionDefinition } from '@votingworks/types';
 import { configureApp } from '../test/helpers/shared_helpers';
 import { scanBallot, withApp } from '../test/helpers/custom_helpers';
 import { PrecinctScannerPollsInfo } from '.';
@@ -161,159 +147,6 @@ test("if there's only one precinct in the election, it's selected automatically 
       precinctId: 'precinct-1',
     });
   });
-});
-
-test('continuous CVR export, including polls closing', async () => {
-  await withApp(
-    {},
-    async ({ apiClient, mockAuth, mockScanner, mockUsbDrive, workspace }) => {
-      await configureApp(apiClient, mockAuth, mockUsbDrive, { testMode: true });
-
-      // Don't wait for continuous export to USB drive in between scans and polls closing so that
-      // we can verify that the continuous export mutex prevents continuous export operations from
-      // interleaving
-      await scanBallot(mockScanner, apiClient, workspace.store, 0, {
-        waitForContinuousExportToUsbDrive: false,
-      });
-      await scanBallot(mockScanner, apiClient, workspace.store, 1, {
-        waitForContinuousExportToUsbDrive: false,
-      });
-      await scanBallot(mockScanner, apiClient, workspace.store, 2, {
-        waitForContinuousExportToUsbDrive: false,
-      });
-
-      expect(
-        await apiClient.exportCastVoteRecordsToUsbDrive({
-          mode: 'polls_closing',
-        })
-      ).toEqual(ok());
-
-      const exportDirectoryPaths = await getCastVoteRecordExportDirectoryPaths(
-        mockUsbDrive.usbDrive
-      );
-      expect(exportDirectoryPaths).toHaveLength(1);
-      const exportDirectoryPath = exportDirectoryPaths[0];
-      expect(exportDirectoryPath).toMatch(/TEST__machine_000__*/);
-
-      const { castVoteRecordExportMetadata, castVoteRecordIterator } = (
-        await readCastVoteRecordExport(exportDirectoryPath)
-      ).unsafeUnwrap();
-
-      expect(castVoteRecordExportMetadata.arePollsClosed).toEqual(true);
-      expect(
-        castVoteRecordExportMetadata.castVoteRecordReportMetadata.vxBatch[0]
-          .NumberSheets
-      ).toEqual(3);
-
-      const castVoteRecords: CVR.CVR[] = (
-        await castVoteRecordIterator.toArray()
-      ).map(
-        (castVoteRecordResult) =>
-          castVoteRecordResult.unsafeUnwrap().castVoteRecord
-      );
-      expect(castVoteRecords).toHaveLength(3);
-
-      for (const castVoteRecord of castVoteRecords) {
-        const tabulationVotes = convertCastVoteRecordVotesToTabulationVotes(
-          assertDefined(getCurrentSnapshot(castVoteRecord))
-        );
-        expect(tabulationVotes).toEqual({
-          attorney: ['john-snow'],
-          'board-of-alderman': [
-            'helen-keller',
-            'steve-jobs',
-            'nikola-tesla',
-            'vincent-van-gogh',
-          ],
-          'chief-of-police': ['natalie-portman'],
-          'city-council': [
-            'marie-curie',
-            'indiana-jones',
-            'mona-lisa',
-            'jackie-chan',
-          ],
-          controller: ['winston-churchill'],
-          mayor: ['sherlock-holmes'],
-          'parks-and-recreation-director': ['charles-darwin'],
-          'public-works-director': ['benjamin-franklin'],
-        });
-      }
-    }
-  );
-});
-
-test('continuous CVR export, including polls closing, followed by a full export', async () => {
-  await withApp(
-    {},
-    async ({ apiClient, mockAuth, mockScanner, mockUsbDrive, workspace }) => {
-      await configureApp(apiClient, mockAuth, mockUsbDrive, { testMode: true });
-
-      // Don't wait for continuous export to USB drive in between scans and polls closing so that
-      // we can verify that the continuous export mutex prevents continuous export operations from
-      // interleaving
-      await scanBallot(mockScanner, apiClient, workspace.store, 0, {
-        waitForContinuousExportToUsbDrive: false,
-      });
-      await scanBallot(mockScanner, apiClient, workspace.store, 1, {
-        waitForContinuousExportToUsbDrive: false,
-      });
-
-      expect(
-        await apiClient.exportCastVoteRecordsToUsbDrive({
-          mode: 'polls_closing',
-        })
-      ).toEqual(ok());
-
-      expect(
-        await apiClient.exportCastVoteRecordsToUsbDrive({
-          mode: 'full_export',
-        })
-      ).toEqual(ok());
-
-      // Expect two export directories, one from continuous export and one from the subsequent full
-      // export. Each should contain two CVRs.
-      const exportDirectoryPaths = await getCastVoteRecordExportDirectoryPaths(
-        mockUsbDrive.usbDrive
-      );
-      expect(exportDirectoryPaths).toHaveLength(2);
-
-      for (const exportDirectoryPath of exportDirectoryPaths) {
-        const { castVoteRecordIterator } = (
-          await readCastVoteRecordExport(exportDirectoryPath)
-        ).unsafeUnwrap();
-        const castVoteRecords: CVR.CVR[] = (
-          await castVoteRecordIterator.toArray()
-        ).map(
-          (castVoteRecordResult) =>
-            castVoteRecordResult.unsafeUnwrap().castVoteRecord
-        );
-        expect(castVoteRecords).toHaveLength(2);
-      }
-    }
-  );
-});
-
-test('CVR resync', async () => {
-  await withApp(
-    {},
-    async ({ apiClient, mockAuth, mockScanner, mockUsbDrive, workspace }) => {
-      await configureApp(apiClient, mockAuth, mockUsbDrive, { testMode: true });
-
-      await scanBallot(mockScanner, apiClient, workspace.store, 0);
-
-      // When a CVR resync is required, the CVR resync modal appears on the "insert your ballot"
-      // screen, i.e. the screen displayed when no card is inserted
-      mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-        Promise.resolve({ status: 'logged_out', reason: 'no_card' })
-      );
-
-      expect(
-        await apiClient.exportCastVoteRecordsToUsbDrive({
-          mode: 'full_export',
-        })
-      ).toEqual(ok());
-    }
-  );
 });
 
 test('setPrecinctSelection will reset polls to closed', async () => {
